Done with the CPU. Yey!
I got lost in control bits while implementing this. At the beginning,
I forgot to consider the type of instruction in implementing the
load bits for the register. But eventually, I was able to figure it
out during debugging. I learned that loading values in the register
and jump operations are only active when the instruction is C.